requirements.txt
pytesseract==0.3.10
openai==0.27.4
opencv-python==4.7.0.72
numpy==1.22.4
Pillow==8.4.0
streamlit==1.20.0